Move RelocateService to data package
