+++ app/src/main/java/org/mozilla/rocket/download/data/RelocateService.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.focus.components;
+package org.mozilla.rocket.download.data;
 
 import android.Manifest;
 import android.app.IntentService;
@@ -30,7 +30,6 @@ import org.mozilla.focus.utils.Constants;
 import org.mozilla.focus.utils.NoRemovableStorageException;
 import org.mozilla.focus.utils.Settings;
 import org.mozilla.focus.utils.StorageUtils;
-import org.mozilla.rocket.download.data.DownloadInfoManager;
 
 import java.io.File;
